third_party/blink/web_tests/platform/win/external/wpt/mathml/presentation-markup/operators/display-operator-min-height-default-font-expected.txt

This is a testharness.js-based test.
[FAIL] Operators using largeop should be bigger
  assert_greater_than: Test ∫ expected a number greater than 20.4 but got 17
Harness: the test ran to completion.
